requirements.txt
flask==3.0.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
colorlog==6.8.0
watchdog==3.0.0
psutil==5.9.6
gitpython==3.1.40 